src/app/pages/home/users-table/users-table.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { UsersTableComponent } from './users-table.component';
import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { Users } from 'src/app/_mock-backend';

describe('UsersTableComponent', () => {
  let component: UsersTableComponent;
  let fixture: ComponentFixture<UsersTableComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [
        HttpClientTestingModule
      ],
      declarations: [ UsersTableComponent ],
      schemas: [CUSTOM_ELEMENTS_SCHEMA]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(UsersTableComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('filterUsers should work', () => {
    component.users = Users;
    component.curPage = 2;
    component.perPage = 5;
    fixture.detectChanges();
    component.filterUsers();
    expect(component.currentUsers.length).toBe(5);
    expect(component.currentUsers[0].id).toBe(component.users[component.perPage * (component.curPage - 1)].id);
  });

});
